requirements.txt
Flask==3.0.3
google-generativeai==0.8.2
gunicorn==22.0.0
python-dotenv==1.0.1
pytest==8.1.1
pytest-flask==1.3.0
pytest-cov==5.0.0
google-cloud-logging==3.9.0
google-cloud-error-reporting==1.11.0
Flask-Limiter==3.7.0
Flask-WTF==1.2.1
Flask-Compress==1.14
Flask-Caching==2.1.0
pydantic==2.6.4
